fix(ui-library): rename tooltip component (#860)
