+++ packages/ui-library/src/components/tooltip/index.ts
@@ -1,13 +1,13 @@
-import { html } from 'lit';
+import { html, nothing } from 'lit';
 import { property, state } from 'lit/decorators.js';
 import { Placement as PlacementType } from '@floating-ui/dom';
+import { classMap } from 'lit/directives/class-map.js';
 import { tooltipPosition } from './tooltip-position';
-import { styleCustom } from './index.css';
+import { styleCustom, dark, light } from './index.css';
 import { TAG_NAME } from './renderFunction';
 
-import { ThemeType } from '../../foundation/_tokens-generated/index.themes';
-import { BlrTooltipBubbleRenderFunction } from '../tooltip-bubble/renderFunction';
 import { LitElementCustom } from '../../utils/lit-element-custom';
+import { ThemeType } from '../../foundation/_tokens-generated/index.themes';
 
 const enterEvents = ['pointerenter', 'focus'];
 const leaveEvents = ['pointerleave', 'blur', 'keydown', 'click'];
@@ -18,7 +18,7 @@ export class BlrTooltip extends LitElementCustom {
   @property() theme?: ThemeType = 'Light';
   @property() message!: string;
   @property() hasArrow?: boolean = true;
-  @property() elevation?: boolean = false;
+  @property() elevation?: boolean = true;
   @property() placement?: PlacementType = 'top';
   @property() offset?: number | string = 4;
 
@@ -35,7 +35,7 @@ export class BlrTooltip extends LitElementCustom {
     const slot = this.renderRoot?.querySelector('slot');
     this._referenceElement = slot?.assignedElements({ flatten: true })[0];
 
-    this._tooltipElement = this.renderRoot?.querySelector('blr-tooltip-bubble');
+    this._tooltipElement = this.renderRoot?.querySelector('#tooltipElement');
 
     if (!this._referenceElement || !this._tooltipElement) {
       return;
@@ -54,14 +54,28 @@ export class BlrTooltip extends LitElementCustom {
   protected hide = () => (this.visible = false);
 
   protected render() {
-    return html` <slot></slot>
-      ${BlrTooltipBubbleRenderFunction({
-        theme: this.theme,
-        message: this.message,
-        hasArrow: this.hasArrow,
-        elevation: this.elevation,
-        visible: this.visible,
-      })}`;
+    const dynamicStyles = this.theme === 'Light' ? [light] : [dark];
+
+    const classes = classMap({
+      [`elevation`]: this.elevation || false,
+      [`visible`]: this.visible || false,
+    });
+    return html`<slot></slot>
+      <div id="tooltipElement">
+        <style>
+          ${dynamicStyles}
+        </style>
+        <div class="${classes}">
+          <div class="content">${this.message}</div>
+          ${this.hasArrow
+            ? html`<div class="arrow">
+                <svg width="12" height="4" fill="none" viewBox="0 0 12 4">
+                  <path d="M6 4C3.738 4 3 0 0 0h12C9 0 8.262 4 6 4Z" />
+                </svg>
+              </div>`
+            : nothing}
+        </div>
+      </div> `;
   }
 }
