+++ packages/ui-library/src/index.ts
@@ -6,8 +6,6 @@ export { BlrButtonText } from './components/button-text';
 // Feedback
 export { BlrLoader } from './components/loader';
 
-export { BlrTooltipBubble } from './components/tooltip-bubble';
-
 export { BlrTooltip } from './components/tooltip';
 
 // Forms
